pyproject.toml
[tool.poetry]
name = "git-archiver"
version = "0.1.0"
description = ""
authors = ["Josh Asplund <joshasplund@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
PyGithub = "^1.55"
pygit2 = "^1.9.1"
GitPython = "^3.1.27"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
git-archiver = 'git_archiver:cli'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
